requirements.txt
requests==2.31.0
pandas==2.2.0
python-Levenshtein==0.25.0
Pyarrow==15.0.0
